+++ STORE/Services/Concrate/ProductSizeService.cs
@@ -1,6 +1,7 @@
 ﻿using STORE.DATA.Repository.Abstract;
 using STORE.DTOs;
 using STORE.ENTITY.Entities;
+using STORE.EXCEPTION;
 using STORE.Services.Abstract;
 using STORE.UnitOfWork.Abstract;
 using System;
@@ -10,10 +11,7 @@ using System.Threading.Tasks;
 
 namespace STORE.Services.Concrate
 {
-    //UnitOfWork yapılacak TODO
-    //Configaration bakılacak
-    //Seed Eklenecekler
-    //DI unutma startupconfiguration
+   
     public class ProductSizeService : IProductSizeService
     {
         private readonly IUnitOfWork _unitOfWork;
@@ -27,6 +25,8 @@ namespace STORE.Services.Concrate
 
         public async Task<ProductSizeDTO> AddProductSizeAsync(ProductSizeDTO productSizeDTO)
         {
+            if (productSizeDTO == null)
+                throw new StoreApiException("Eksik yada hatalı bilgi girişi yaptınız");
             var productSize = new ProductSize
             {
                 Name = productSizeDTO.Name
@@ -40,6 +40,8 @@ namespace STORE.Services.Concrate
 
         public async Task DeleteProductSizeAsync(int productSizeId)
         {
+            if (productSizeId <= 0)
+                throw new StoreApiException("Silmek istedğiniz beden bilgisine erişilemedi");
             await _productSizeRepository.DeleteAsync(productSizeId).ConfigureAwait(false);
             await _unitOfWork.SaveChangeAsync().ConfigureAwait(false);
         }
@@ -47,7 +49,8 @@ namespace STORE.Services.Concrate
         public async Task<List<ProductSizeDTO>> GetAllProductSizesAsync()
         {
             var productSizes = await _productSizeRepository.GetAllAsync().ConfigureAwait(false);
-
+            if (productSizes == null)
+                throw new StoreApiException("Kayıtlı herhangi bir beden bulunamadı");
             var productSizeDTOs = productSizes != null ?
                                   (from ps in productSizes
                                    select new ProductSizeDTO
@@ -63,6 +66,8 @@ namespace STORE.Services.Concrate
         public async Task<ProductSizeDTO> GetByIdProductSizeAsync(int productSizeId)
         {
             var productSize = await _productSizeRepository.GetByIdAsync(productSizeId).ConfigureAwait(false);
+            if (productSize == null)
+                throw new StoreApiException("İstenilen beden bilgisine erişilemedi");
             var productSizeDTO = new ProductSizeDTO
             {
                 Id = productSize.Id,
@@ -77,7 +82,8 @@ namespace STORE.Services.Concrate
         public async Task<ProductSizeDTO> UpdateProcductSizezAsync(ProductSizeDTO productSizeDTO)
         {
             var size = await _productSizeRepository.GetByIdAsync(productSizeDTO.Id).ConfigureAwait(false);
-
+            if (size == null)
+                throw new StoreApiException("Güncellenmek istenen bedene erişilemedi");
             size.Name = productSizeDTO.Name;
            /* var productSize = new ProductSize
             {
